:sparkles: Finished working on the validations of the login and singup page
