+++ src/redux/user/user.js
@@ -1,4 +1,4 @@
-import { setLocalStorage } from '../../logics/localStore';
+// import { setLocalStorage } from '../../logics/localStore';
 import userURL from '../../logics/urls';
 
 const FETCH_USERS = 'users/FETCH_USERS';
@@ -20,24 +20,6 @@ export const fetchUsers = () => (dispatch) => {
   fetch(userURL).then((res) => res.json()).then((data) => {
     dispatch(getUsers(data.data));
   });
-};
-
-export const signup = (user) => (dispatch) => {
-  fetch(userURL, {
-    method: 'POST',
-    headers: {
-      'Content-type': 'application/json; charset=UTF-8',
-    },
-    body: JSON.stringify(user),
-  })
-    .then((response) => response.json())
-    .then((data) => {
-      if (data.status === 'SUCCESS') {
-        dispatch(postUser(user));
-        const localData = { user_id: user.id, loggedIn: true };
-        setLocalStorage(localData);
-      }
-    });
 };
 
 const userReducer = (state = initialState, action) => {
